Fixing Issues #248, #257
* Can not create thold for graph
* Can not add a threshold with threshold template, can not choose device

This change will remove the Template dropdown in cases where there are no matching Templates that match the Graph.

Even if there are Templates that match the graph, if the Thold for the Template is removed, then the option for that Template will also be removed.

It also removes a large amount of redundant codes from the Wizard creation process.

Lastly, there was a typo including the word 'NOT' when it should not, and some issues with WC3 validation that have been corrected.